+++ types/beatmap/v3/customEvent.ts
@@ -1,5 +1,73 @@
-import { IHeckCustomEvent } from './heck.ts';
-import { IChromaCustomEvent } from './chroma.ts';
-import { INECustomEvent } from './noodleExtensions.ts';
+import {
+    IChromaCustomEventDataAnimateComponent,
+    IChromaCustomEventDataAnimateTrack,
+    IChromaCustomEventDataAssignFogTrack,
+    IChromaCustomEventDataAssignPathAnimation,
+} from './chroma.ts';
+import { IHeckCustomEventDataAnimateTrack, IHeckCustomEventDataAssignPathAnimation } from './heck.ts';
+import {
+    INECustomEventDataAnimateTrack,
+    INECustomEventDataAssignPathAnimation,
+    INECustomEventDataAssignPlayerToTrack,
+    INECustomEventDataAssignTrackParent,
+} from './noodleExtensions.ts';
 
-export type ICustomEvent = IHeckCustomEvent | IChromaCustomEvent | INECustomEvent;
+export type ICustomEventDataAnimateTrack =
+    & IHeckCustomEventDataAnimateTrack
+    & IChromaCustomEventDataAnimateTrack
+    & INECustomEventDataAnimateTrack;
+
+export type ICustomEventDataAssignPathAnimation =
+    & IHeckCustomEventDataAssignPathAnimation
+    & IChromaCustomEventDataAssignPathAnimation
+    & INECustomEventDataAssignPathAnimation;
+
+/** Custom Event interface for AnimateTrack. */
+export interface ICustomEventAnimateTrack {
+    b: number;
+    t: 'AnimateTrack';
+    d: ICustomEventDataAnimateTrack;
+}
+
+/** Custom Event interface for AssignPathAnimation. */
+export interface ICustomEventAssignPathAnimation {
+    b: number;
+    t: 'AssignPathAnimation';
+    d: ICustomEventDataAssignPathAnimation;
+}
+
+/** Custom Event interface for AssignFogTrack. */
+export interface ICustomEventAssignFogTrack {
+    b: number;
+    t: 'AssignFogTrack';
+    d: IChromaCustomEventDataAssignFogTrack;
+}
+
+/** Custom Event interface for AnimateComponent. */
+export interface ICustomEventAnimateComponent {
+    b: number;
+    t: 'AnimateComponent';
+    d: IChromaCustomEventDataAnimateComponent;
+}
+
+/** Custom Event interface for AssignTrackParent. */
+export interface ICustomEventAssignTrackParent {
+    b: number;
+    t: 'AssignTrackParent';
+    d: INECustomEventDataAssignTrackParent;
+}
+
+/** Custom Event interface for AssignPlayerToTrack. */
+export interface ICustomEventAssignPlayerToTrack {
+    b: number;
+    t: 'AssignPlayerToTrack';
+    d: INECustomEventDataAssignPlayerToTrack;
+}
+
+export type ICustomEvent =
+    | ICustomEventAnimateTrack
+    | ICustomEventAssignPathAnimation
+    | ICustomEventAssignFogTrack
+    | ICustomEventAnimateComponent
+    | ICustomEventAssignTrackParent
+    | ICustomEventAssignPlayerToTrack;